requirements.txt
# crawling
requests
beautifulsoup4
PyYAML
pydantic
pandas